requirements.txt
aiohttp==3.5.4
altgraph==0.17
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.4
et-xmlfile==1.0.1
future==0.18.2
google-api-core==1.16.0
google-auth==1.11.2
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
google-cloud-translate==2.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
multidict==4.7.4
numpy==1.18.1
openpyxl==3.0.3
pandas==0.25.3
pdfminer.six==20200104
pefile==2019.4.18
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycryptodome==3.9.4
PyInstaller==3.6
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
rsa==4.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
